+++ modules/webtools/src/cpp/RepositoryCheckout.cpp
@@ -111,13 +111,13 @@ RepositoryCreateAndCheckoutBranch(
 }
 //=============================================================================
 static void
-RepositoryCheckout(const std::wstring& localPath, const std::wstring& branchName, bool detach,
+RepositoryCheckoutDetached(const std::wstring& localPath, const std::wstring& sha1OrTag,
     std::wstring& errorMessage)
 {
     git_libgit2_init();
     git_repository* repo = NULL;
     std::string localPathUtf8 = wstring_to_utf8(localPath);
-    std::string branchNameUtf8 = wstring_to_utf8(branchName);
+    std::string sha1OrTagUtf8 = wstring_to_utf8(sha1OrTag);
 
     int errorCode = git_repository_open(&repo, localPathUtf8.c_str());
     if (errorCode != 0) {
@@ -143,10 +143,10 @@ RepositoryCheckout(const std::wstring& localPath, const std::wstring& branchName
         return;
     }
     git_object* treeish = NULL;
-    errorCode = git_revparse_single(&treeish, repo, branchNameUtf8.c_str());
+    errorCode = git_revparse_single(&treeish, repo, sha1OrTagUtf8.c_str());
     if (errorCode != 0) {
-        std::string branchTempName = std::string("origin/") + branchNameUtf8;
-        errorCode = git_revparse_single(&treeish, repo, branchTempName.c_str());
+        std::string sha1OrTagTempName = std::string("origin/") + sha1OrTagUtf8;
+        errorCode = git_revparse_single(&treeish, repo, sha1OrTagTempName.c_str());
     }
     if (errorCode != 0) {
         errorMessage = gitErrorCodeToMessage(errorCode);
@@ -185,11 +185,11 @@ RepositoryCheckout(
     } else {
         bool isTag = RepositoryIsTag(localPath, branchOrTag);
         if (isTag) {
-            RepositoryCheckout(localPath, branchOrTag, errorMessage);
+            RepositoryCheckoutDetached(localPath, branchOrTag, errorMessage);
         } else {
             bool isSHA1 = RepositoryIsSHA1(localPath, branchOrTag);
             if (isSHA1) {
-                RepositoryCheckout(localPath, branchOrTag, errorMessage);
+                RepositoryCheckoutDetached(localPath, branchOrTag, errorMessage);
             } else {
                 errorMessage = _W("Valid tag or branch name expected.");
             }